+++ python/oneflow/utils/data/__init__.py
@@ -32,8 +32,6 @@ from oneflow.utils.data.dataset import IterableDataset as IterDataPipe
 from oneflow.utils.data.dataloader import (
     DataLoader,
     _DatasetKind,
-    KeyErrorMessage,
-    ExceptionWrapper,
 )
 from oneflow.utils.data.decorator import (
     functional_datapipe,
