+++ python/oneflow/utils/data/_utils/pin_memory.py
@@ -26,7 +26,7 @@ import collections.abc
 import queue
 
 from . import MP_STATUS_CHECK_INTERVAL
-from oneflow.utils.data import ExceptionWrapper
+from oneflow._utils import ExceptionWrapper
 
 container_abcs = collections.abc
 string_classes = (str, bytes)
