+++ python/oneflow/utils/data/dataloader.py
@@ -25,47 +25,8 @@ from typing import Any, Callable, TypeVar, Generic, Sequence, List, Optional
 import multiprocessing as python_multiprocessing
 
 import oneflow.multiprocessing as multiprocessing
+from oneflow._utils import ExceptionWrapper
 import oneflow as flow
-from oneflow.utils.data import _utils
-
-
-class KeyErrorMessage(str):
-    r"""str subclass that returns itself in repr"""
-
-    def __repr__(self):
-        return self
-
-
-class ExceptionWrapper(object):
-    r"""Wraps an exception plus traceback to communicate across threads"""
-
-    def __init__(self, exc_info=None, where="in background"):
-        # It is important that we don't store exc_info, see
-        # NOTE [ Python Traceback Reference Cycle Problem ]
-        if exc_info is None:
-            exc_info = sys.exc_info()
-        self.exc_type = exc_info[0]
-        self.exc_msg = "".join(traceback.format_exception(*exc_info))
-        self.where = where
-
-    def reraise(self):
-        r"""Reraises the wrapped exception in the current thread"""
-        # Format a message such as: "Caught ValueError in DataLoader worker
-        # process 2. Original Traceback:", followed by the traceback.
-        msg = "Caught {} {}.\nOriginal {}".format(
-            self.exc_type.__name__, self.where, self.exc_msg
-        )
-        if self.exc_type == KeyError:
-            # KeyError calls repr() on its argument (usually a dict key). This
-            # makes stack traces unreadable. It will not be changed in Python
-            # (https://bugs.python.org/issue2651), so we work around it.
-            msg = KeyErrorMessage(msg)
-        elif getattr(self.exc_type, "message", None):
-            # Some exceptions have first argument as non-str but explicitly
-            # have message field
-            raise self.exc_type(message=msg)
-        raise self.exc_type(msg)
-
 
 string_classes = (str, bytes)
